+++ src/main/java/org/aksw/gerbil/transfer/nif/Relation.java
@@ -1,10 +1,21 @@
 package org.aksw.gerbil.transfer.nif;
 
-import org.apache.jena.graph.Triple;
+import java.util.Set;
 
+
+/**
+ * Implements a Relation Object 
+ *
+ */
 public interface Relation extends Marking {
 
-    public void setRelation(Triple relation);
+    public void setRelation(String subject, String predicate, String object);
     
-    public Triple getRelation();
+    public Set<String>[] getRelation();
+    
+    public Set<String> getSubject();
+    
+    public Set<String> getPredicate();
+    
+    public Set<String> getObject();
 }
