+++ src/main/java/org/aksw/gerbil/transfer/nif/data/RelationImpl.java
@@ -1,59 +1,64 @@
 package org.aksw.gerbil.transfer.nif.data;
 
+import java.util.HashSet;
+import java.util.Set;
+
 import org.aksw.gerbil.transfer.nif.Relation;
-import org.apache.jena.graph.Triple;
 
 public class RelationImpl extends AbstractMarkingImpl implements Relation {
 
-    protected Triple relation;
+	private Set<String> subjects = new HashSet<String>();
+	private Set<String> predicates = new HashSet<String>();
+	private Set<String> objects = new HashSet<String>();
+	
+	public RelationImpl() {
+		
+	}
+	
+	public RelationImpl(String subject, String predicate, String object) {
+		this.setRelation(subject, predicate, object);
+	}
+	
+	public RelationImpl(Set<String> subjects, Set<String> predicates, Set<String> objects) {
+		this.subjects.addAll(subjects);
+		this.predicates.addAll(predicates);
+		this.objects.addAll(objects);
+	}
 
-    public RelationImpl(Triple relation) {
-        setRelation(relation);
-    }
+	@Override
+	public void setRelation(String subject, String predicate, String object) {
+		this.subjects.add(subject);
+		this.predicates.add(predicate);
+		this.objects.add(object);
+	}
 
-    public RelationImpl(Relation relation) {
-        this(relation.getRelation());
-    }
+	@Override
+	public Set<String>[] getRelation() {
+		Set<String>[] relation = new Set[3];
+		relation[0]=subjects;
+		relation[1]=predicates;
+		relation[2]=objects;
+		return relation;
+	}
 
-    @Override
-    public void setRelation(Triple relation) {
-        this.relation = relation;
-    }
+	@Override
+	public Set<String> getSubject() {
+		return subjects;
+	}
 
-    @Override
-    public Triple getRelation() {
-        return relation;
-    }
+	@Override
+	public Set<String> getPredicate() {
+		return predicates;
+	}
 
-    @Override
-    public String toString() {
-        return relation.toString();
-    }
+	@Override
+	public Set<String> getObject() {
+		return objects;
+	}
 
-    @Override
-    public Object clone() throws CloneNotSupportedException {
-        return new RelationImpl(this);
-    }
+	@Override
+	public Object clone() throws CloneNotSupportedException {
+		return new RelationImpl(subjects, predicates, objects);
+	}
 
-    @Override
-    public int hashCode() {
-        return ((relation == null) ? 0 : relation.hashCode());
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        RelationImpl other = (RelationImpl) obj;
-        if (relation == null) {
-            if (other.relation != null)
-                return false;
-        } else if (!relation.equals(other.relation))
-            return false;
-        return true;
-    }
 }
